Inserido AutoMapper no projeto UI Consertado tipos nas Entidades Feito funcionalidade de alteração de senha

+++ DEKL.CP.Domain/Entities/Banco.cs
@@ -5,8 +5,8 @@
         public string Nome { get; set; }
         public string Telefone { get; set; }
         public string Email { get; set; }
-        public double TaxaChequeEspecial { get; set; }
-        public double TaxaEmprestimo { get; set; }
+        public decimal TaxaChequeEspecial { get; set; }
+        public decimal TaxaEmprestimo { get; set; }
         public double EncargResp { get; set; }
         public int? EnderecoId { get; set; }
         public virtual Endereco Endereco { get; set; }
